// ISAFront-master/src/app/model/Patient.ts
export class Patient {
    id: number;
    username: string = "";
    password: string = "";
    repassword: string = "";
    firstName: string = "";
    lastName: string = "";
    email: string = "";
    address: string = "";
    city: string = "";
    country: string = "";
    mobileNumber: number;
    jmbg: number;
}